backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.9.2
email-validator==2.2.0
argon2-cffi==23.1.0
pydantic-settings==2.1.0
